+++ examples/advanced/interpolate_scalar5.py
@@ -15,7 +15,7 @@ surf = Grid(res=[25,25])
 
 # Pick n random points on the surface
 ids = np.random.randint(0, surf.npoints, 10)
-pts = surf.points(ids)
+pts = surf.vertices[ids]
 
 # Create a set of random scalars
 scals1 = np.random.randn(10) * 0.1 
@@ -32,7 +32,7 @@ pts2 = Points(ptss2, r=15, c="purple5")
 
 # Warp the surface to match the interpolated points
 ptsource, pttarget = [], []
-for pt in pts2.points():
+for pt in pts2.vertices:
     pt_surf = surf.closest_point(pt)
     ptsource.append(pt_surf)
     pttarget.append(pt)