+++ examples/other/napari1.py
@@ -6,9 +6,9 @@ import vedo
 surf = vedo.Mesh(vedo.dataurl+"beethoven.ply").triangulate().compute_normals()
 surf.rotate_x(180).rotate_y(60)
 
-vertices = surf.points()
-faces    = np.array(surf.cells)
-normals  = surf.normals()
+vertices = surf.vertices
+faces    = surf.cells
+normals  = surf.vertex_normals
 # generate vertex values by projecting normals on a "lighting vector"
 values   = np.dot(normals, [-1, 1, 1])